Add fragment option

+++ Tests/StaticContentGeneratorTests.swift
@@ -17,6 +17,7 @@ final class StaticContentGeneratorTests: XCTestCase {
     private static let testPost = Post(slug: "example-post", title: "Example Post", topics: [], createdDate: Date(), publishStatus: .public)
     
     override func setUpWithError() throws {
+        try deleteExistingContents()
         try FileManager.default.createDirectory(at: StaticContentGeneratorTests.testPostDirectoryURL, withIntermediateDirectories: true)
         guard let exampleContentFileURL = Bundle.module.url(forResource: "Example", withExtension: "md") else {
             throw CocoaError(.fileNoSuchFile)
@@ -25,17 +26,23 @@ final class StaticContentGeneratorTests: XCTestCase {
     }
 
     override func tearDownWithError() throws {
-        try FileManager.default.removeItem(at: StaticContentGeneratorTests.testPostDirectoryURL)
+        try deleteExistingContents()
     }
     
-    func testGeneratingMarkupFile() throws {
+    func deleteExistingContents() throws {
+        if FileManager.default.fileExists(atPath: StaticContentGeneratorTests.testPostDirectoryURL.path) {
+            try FileManager.default.removeItem(at: StaticContentGeneratorTests.testPostDirectoryURL)
+        }
+    }
+    
+    func testGeneratingStaticFragment() throws {
         let generator = StaticContentGenerator(contentDirectory: StaticContentGeneratorTests.testPostDirectoryURL)
         let markdownFile = try MarkdownFile(fileURL: StaticContentGeneratorTests.testPostDirectoryURL.appending(path: "Example.md"))
-        let generatedMarkup = markdownFile.markupRepresentation(strippingFirstHeadingElement: true)
+        let generatedMarkup = markdownFile.markupRepresentation(fragment: true)
         
         self.measure {
             do {
-                try generator.generateStaticContent(for: StaticContentGeneratorTests.testPost, with: markdownFile)
+                try generator.generateStaticContent(for: StaticContentGeneratorTests.testPost, with: markdownFile, fragment: true)
                 
                 let generatedFileURL = StaticContentGeneratorTests.testPostDirectoryURL.appending(path: "index.html")
                 let doesGeneratedFileExist = FileManager.default.fileExists(atPath: generatedFileURL.path)
@@ -43,6 +50,31 @@ final class StaticContentGeneratorTests: XCTestCase {
                 
                 let generatedFileContent = try String(contentsOf: generatedFileURL)
                 assert(generatedFileContent == generatedMarkup)
+                assert(!generatedFileContent.contains("<html>"))
+            }
+            catch {
+                XCTFail("Error generating static content: \(error)")
+            }
+        }
+    }
+    
+    func testGeneratingStaticFullyFormed() throws {
+        let generator = StaticContentGenerator(contentDirectory: StaticContentGeneratorTests.testPostDirectoryURL)
+        let markdownFile = try MarkdownFile(fileURL: StaticContentGeneratorTests.testPostDirectoryURL.appending(path: "Example.md"))
+        let generatedMarkup = markdownFile.markupRepresentation(fragment: false)
+        
+        self.measure {
+            do {
+                try generator.generateStaticContent(for: StaticContentGeneratorTests.testPost, with: markdownFile, fragment: false)
+                
+                let generatedFileURL = StaticContentGeneratorTests.testPostDirectoryURL.appending(path: "index.html")
+                let doesGeneratedFileExist = FileManager.default.fileExists(atPath: generatedFileURL.path)
+                assert(doesGeneratedFileExist)
+                
+                let generatedFileContent = try String(contentsOf: generatedFileURL)
+                assert(generatedFileContent == generatedMarkup)
+                assert(generatedFileContent.contains("<html>"))
+                assert(generatedFileContent.contains("<title>"))
             }
             catch {
                 XCTFail("Error generating static content: \(error)")
